wsp: Move a majority of the constants to a header file.
